requirements.txt
appdirs==1.4.3
Django==1.11
djangorestframework==3.6.2
packaging==16.8
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
uWSGI==2.0.15
psycopg2==2.7.3.2
